requirements.txt
alembic==1.0.5
aniso8601==4.1.0
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-JWT-Extended==3.15.0
Flask-Migrate==2.3.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
more-itertools==5.0.0
passlib==1.7.1
pluggy==0.8.1
py==1.7.0
PyJWT==1.7.1
pytest==4.1.1
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
pytz==2018.9
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.16
urllib3==1.24.1
Werkzeug==0.14.1
